Move sleep_min code to main

+++ philo/main.c
@@ -6,7 +6,7 @@
 /*   By: lyanga <lyanga@student.42singapore.sg>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/21 01:59:17 by lyanga            #+#    #+#             */
-/*   Updated: 2026/01/06 17:02:48 by lyanga           ###   ########.fr       */
+/*   Updated: 2026/01/06 18:15:01 by lyanga           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -138,7 +138,7 @@ static int init_setup(struct sim_info *info, struct philosopher **philosophers, 
 // bufferUS is a constant buffer time in microseconds that is meant to give the program leeway time to act
 #define BUFFERUS 100
 
-void sleep_for(uint64_t sleepeatdiff, uint64_t timeleft, uint64_t percentage, uint64_t ttd)
+void sleep_for(int sleepeatdiff, uint64_t timeleft, uint64_t percentage, uint64_t ttd)
 {
     // if time to sleep is shorter than time to eat, minimum time to usleep should be the diff + BUFFERUS
     if (sleepeatdiff > 0)
